Start of voxel octree

+++ src/TerrainGeneration.Application.VoxelOctree/OctreeEventQueue/OctreeEventQueue.cs
@@ -13,7 +13,7 @@ namespace TerrainGeneration.Application.VoxelOctree.OctreeEventQueue
     {
         public Queue<IOctreeEvent> EventQueue;
         public IRenderOctree EventTargetTree;
-        public int WorkBudget;
+        public uint WorkBudget;
 
         /// <summary>
         /// Creates an octree event queue. Processes the events sent from the abstract octree against the render octree.
@@ -22,7 +22,7 @@ namespace TerrainGeneration.Application.VoxelOctree.OctreeEventQueue
         /// <param name="workBudget"></param>
         /// <exception cref="ArgumentNullException"></exception>
         /// <exception cref="ArgumentException"></exception>
-        public OctreeEventQueue(IRenderOctree eventTargetTree, int workBudget) 
+        public OctreeEventQueue(IRenderOctree eventTargetTree, uint workBudget) 
         { 
             if (eventTargetTree== null)
             {
@@ -53,7 +53,7 @@ namespace TerrainGeneration.Application.VoxelOctree.OctreeEventQueue
         /// </summary>
         public void Process()
         {
-            int numWork = Math.Min(WorkBudget, EventQueue.Count);
+            uint numWork = Math.Min(WorkBudget, (uint)EventQueue.Count);
             IOctreeEvent[] eventsToProcess = new IOctreeEvent[numWork];
 
             // Grab events to process this frame
